Extract kubelet network code into its own file
